Cloud func and misc cleanup (#26)
Note: the deploy preview here won't work since it will be hitting the
live cloud functions which are not updated with the changes in this PR
yet. If you want to test just the frontend on the Netlify preview, put
`?mock=true` at the end of the url and it will use fake data instead of
a real api.

- convert cloud funcs and consuming frontend api funcs to use
json-encoded body params instead of url params
- make some frontend types more specific
- fix button component unique key warning
- remove vestigial "icon" className in a few places
- display summary of inputs on results page
- add per-species example genes
- make cloud functions use POST and request body for passing params
- return appropriate status codes and error messages from cloud funcs
- move documentation of cloud funcs to separate readme for readability
(headings to split up/organize content, code blocks for syntax
highlighting, etc). do a bit of rewriting of comments for clarity.
- rename "run_pipeline" to "ml" for consistency

---------

Co-authored-by: Christopher Andrew Mancuso <chrismancuso1984@gmail.com>