Changes from PyGenePlexus (#45)
The items in here that changed are from commits 5722d9b, d07ce39 and
62cfc9a in the PyGenePlexus repo. Things that specifically changed that
might need a UI tweak here
- In both the probability and similarity tables there are columns from
z-scores and p-values
- The edge lists are now weighted again
- using Mondo instead of DisGeNet for disease-gene annotations
- In similarity table there is now a column called "Task" that describes
the type of GSC it is from
- changed how the auPRC values are returned, especially when they can't
be calculated
- Added a column "Gene Name" to the tables the show how the input IDs
are converted
- If training and showing results for different species, the probability
data frame now can have known-novel and Class-Label columns and this
information is based on one-to-one ortholog info. Can be used in network
figure.

Other UI ask that could be good in this PR is possibly showing only
StTRING as an option and then having a button a user clicks labeled
"More Options" that would show the BioGRID and IMP networks. Similarity
the GSC could be default to Combined and the user needs to click a
button to see the individual GSCs. This is #44

I also added a script to gather and zip the data for the function. I
uploaded the new data to the GCP bucket.

---------

Co-authored-by: Vincent Rubinetti <vince.rubinetti@gmail.com>