Rearrange new analysis page (#52)
Based on Chris's sketch here:


[20240924_UI_thoughts[12].pdf](https://github.com/user-attachments/files/17604167/20240924_UI_thoughts.12.pdf)

I have made tweaks though, because the wording and order of things
needed to be changed. For example, if we're going to put the negatives
in the options section, the "Check genes" section needs to come after it
because it also checks those. Generally, it is clearer if something the
user inputs (negatives) doesn't affect anything that came before it
(check genes results).

- fix css z-index bug in toasts/view corner
- add another summarizing sentence to home page
- new analysis page changes 
- make "inputType" case insensitive (backend/package seems to normalize
input genes to uppercase). make more succinct by just using `+` and `-`
chars
- move species selects to first section, side by side, with arrow in
between
- remove "show negatives" checkbox and move negatives text box to
parameters section
- rename "choose parameters" section to "options", and wrap in
collapsible
- rename "check genes" section to "pre-check" genes to make it more
clear at a glance what it's for. hide paragraph of explanatory text in
tooltip for clean-ness.