Move theme vars from css to js (#59)
We were experiencing an intermittent problem on Safari where the Network
component would be all black:


![image](https://github.com/user-attachments/assets/021cfdaa-a235-4de3-b017-88d4ebd9f5c4)

Looking at the code, I have to assume that this is due to the current
`theme()` function returning nothing due to some race condition where
the `/global/theme.css` CSS vars are not present before that function is
run (multiple times) at the top of `/components/Network.tsx`. My
understanding was that, since we were importing the global CSS before
anything else, and since Vite injects it as a normal style block in the
document head, it should've always been defined before any of the JS
that depended on it ran. But I guess that is not always the case;
perhaps only in Safari.


This PR moves declarations of theme variables from the CSS to JS, to
make sure they are always defined and available when being imported in
JS. This was simpler, in the case of this app, than e.g. making a
reactive hook that actually waits for the page to be fully rendered (as
it works in the sister molevolvr 2.0 repo). It also has the side benefit
of TypeScript now knowing exactly which theme variables are available to
be used.